entity.getcomponent now returns optional. all angles are now in degrees. most doubles are now floats. add spaceship image. refactor rendersystem again, simplifying greatly. rethink key event polling.

+++ CoinModule/src/main/java/io/asteroidsjaylib/coin/CoinEntity.java
@@ -29,8 +29,8 @@ public class CoinEntity extends BaseEntity {
 
         String text = value+"";
         TextComponent textComponent = new TextComponent(text, 12, BLACK);
-        textComponent.xoffset = -(double) MeasureText(text, 12) / 2;
-        textComponent.yoffset = -12.0 / 2;
+        textComponent.xoffset = (float) -MeasureText(text, 12) / 2;
+        textComponent.yoffset = (float) (-12.0 / 2);
         renderTag.addRenderComponent(textComponent, 1);
 
         this.addComponent(renderTag);
