+++ CommonModule/src/main/java/io/asteroidsjaylib/common/ecs/IteratingSystem.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public abstract class IteratingSystem extends BaseSystem {
     @Override
-    public void update(World world, List<BaseEntity> entities, double deltaTime){
+    public void update(World world, List<BaseEntity> entities, float deltaTime){
         for(BaseEntity entity : entities){
             processEntity(world, entity, deltaTime);
         }
     }
-    public abstract void processEntity(World world, BaseEntity entity, double deltaTime);
+    public abstract void processEntity(World world, BaseEntity entity, float deltaTime);
 }
